fix(storage): disable default KV cache and split durable modules
Keep WAL/cache semantics. Default KvReadCacheConfig retains nothing
until a reserved or governed budget is supplied. Split oversized
durable/mod.rs and group.rs into named modules without comment-stripping
or include! shards. Disclose ReadyKvRoot, current_kv_root_if_ready, and
FaultPoint::AfterWalPublication. Add volume-backed WAL crash/reopen,
torn-tail, and disabled-policy replay regressions.
