+++ crates/astrid-storage/src/engine/durable/mod.rs
@@ -7,17 +7,16 @@
 
 use std::collections::{BTreeMap, BTreeSet};
 use std::fmt;
-use std::fs::File as NativeFile;
 #[cfg(test)]
 use std::fs::OpenOptions;
-use std::io::{self, BufWriter, Read, Seek, SeekFrom, Write};
+use std::io::{self, BufWriter, Seek, SeekFrom};
 use std::num::{NonZeroU32, NonZeroU64};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::sync::atomic::AtomicU8;
 use std::time::Duration;
 
-use crate::volume::{AstridVolume, VolumeFile, VolumeMetadata, VolumeRegion};
+use crate::volume::{AstridVolume, VolumeRegion};
 
 use crate::content_dag::ContentError;
 use crate::storage_model::{
@@ -703,193 +702,8 @@ impl<P: Ord, I, C> Drop for DurableEngine<P, I, C> {
     }
 }
 
-#[derive(Debug)]
-enum StoreLock {
-    Native(NativeFile),
-    Volume,
-}
-
-/// Byte-stream handle used by the durable engine.
-///
-/// Native files remain available for legacy import and format tests. Runtime
-/// layout two opens the same engine over [`VolumeFile`], so object and root
-/// authority no longer depends on a host directory namespace.
-#[derive(Debug)]
-enum File {
-    Native(NativeFile),
-    Volume(VolumeFile),
-}
-
-/// Common random-access durable byte stream used by both legacy host files and
-/// Astrid volume regions.
-///
-/// Keeping framing generic over this interface is important: the frozen store
-/// format is independent of whichever media provider owns the bytes.
-trait DurableIo: Read + Write + Seek {
-    fn durable_metadata(&self) -> io::Result<StoreMetadata>;
-    fn durable_set_len(&self, length: u64) -> io::Result<()>;
-    fn durable_sync_data(&self) -> io::Result<()>;
-    fn durable_read_at(&self, buffer: &mut [u8], offset: u64) -> io::Result<usize>;
-}
-
-impl File {
-    fn native(file: NativeFile) -> Self {
-        Self::Native(file)
-    }
-
-    fn volume(
-        volume: Arc<dyn AstridVolume>,
-        name: &str,
-        create: bool,
-    ) -> Result<Self, DurableError> {
-        let region =
-            VolumeRegion::new(name).map_err(|source| io_error("validate volume region", source))?;
-        VolumeFile::open(volume, region, create)
-            .map(Self::Volume)
-            .map_err(|source| io_error("open volume region", source))
-    }
-
-    fn try_clone(&self) -> io::Result<Self> {
-        match self {
-            Self::Native(file) => file.try_clone().map(Self::Native),
-            Self::Volume(file) => file.try_clone().map(Self::Volume),
-        }
-    }
-
-    fn sync_data(&self) -> io::Result<()> {
-        match self {
-            Self::Native(file) => file.sync_data(),
-            Self::Volume(file) => file.sync_data(),
-        }
-    }
-
-    fn set_len(&self, length: u64) -> io::Result<()> {
-        match self {
-            Self::Native(file) => file.set_len(length),
-            Self::Volume(file) => file.set_len(length),
-        }
-    }
-
-    fn metadata(&self) -> io::Result<StoreMetadata> {
-        match self {
-            Self::Native(file) => file.metadata().map(StoreMetadata::Native),
-            Self::Volume(file) => file.metadata().map(StoreMetadata::Volume),
-        }
-    }
-
-    fn read_at(&self, buffer: &mut [u8], offset: u64) -> io::Result<usize> {
-        match self {
-            #[cfg(unix)]
-            Self::Native(file) => std::os::unix::fs::FileExt::read_at(file, buffer, offset),
-            #[cfg(windows)]
-            Self::Native(file) => std::os::windows::fs::FileExt::seek_read(file, buffer, offset),
-            #[cfg(not(any(unix, windows)))]
-            Self::Native(file) => {
-                let mut reader = file.try_clone()?;
-                reader.seek(SeekFrom::Start(offset))?;
-                reader.read(buffer)
-            },
-            Self::Volume(file) => file.read_at(buffer, offset),
-        }
-    }
-}
-
-impl std::io::Read for File {
-    fn read(&mut self, buffer: &mut [u8]) -> io::Result<usize> {
-        match self {
-            Self::Native(file) => file.read(buffer),
-            Self::Volume(file) => file.read(buffer),
-        }
-    }
-}
-
-impl std::io::Write for File {
-    fn write(&mut self, bytes: &[u8]) -> io::Result<usize> {
-        match self {
-            Self::Native(file) => file.write(bytes),
-            Self::Volume(file) => file.write(bytes),
-        }
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        match self {
-            Self::Native(file) => file.flush(),
-            Self::Volume(file) => file.flush(),
-        }
-    }
-}
-
-impl std::io::Seek for File {
-    fn seek(&mut self, position: SeekFrom) -> io::Result<u64> {
-        match self {
-            Self::Native(file) => file.seek(position),
-            Self::Volume(file) => file.seek(position),
-        }
-    }
-}
-
-impl DurableIo for File {
-    fn durable_metadata(&self) -> io::Result<StoreMetadata> {
-        self.metadata()
-    }
-
-    fn durable_set_len(&self, length: u64) -> io::Result<()> {
-        self.set_len(length)
-    }
-
-    fn durable_sync_data(&self) -> io::Result<()> {
-        self.sync_data()
-    }
-
-    fn durable_read_at(&self, buffer: &mut [u8], offset: u64) -> io::Result<usize> {
-        self.read_at(buffer, offset)
-    }
-}
-
-impl DurableIo for NativeFile {
-    fn durable_metadata(&self) -> io::Result<StoreMetadata> {
-        self.metadata().map(StoreMetadata::Native)
-    }
-
-    fn durable_set_len(&self, length: u64) -> io::Result<()> {
-        self.set_len(length)
-    }
-
-    fn durable_sync_data(&self) -> io::Result<()> {
-        self.sync_data()
-    }
-
-    fn durable_read_at(&self, buffer: &mut [u8], offset: u64) -> io::Result<usize> {
-        #[cfg(unix)]
-        {
-            std::os::unix::fs::FileExt::read_at(self, buffer, offset)
-        }
-        #[cfg(windows)]
-        {
-            std::os::windows::fs::FileExt::seek_read(self, buffer, offset)
-        }
-        #[cfg(not(any(unix, windows)))]
-        {
-            let mut reader = self.try_clone()?;
-            reader.seek(SeekFrom::Start(offset))?;
-            reader.read(buffer)
-        }
-    }
-}
-
-enum StoreMetadata {
-    Native(std::fs::Metadata),
-    Volume(VolumeMetadata),
-}
-
-impl StoreMetadata {
-    fn len(&self) -> u64 {
-        match self {
-            Self::Native(metadata) => metadata.len(),
-            Self::Volume(metadata) => metadata.len(),
-        }
-    }
-}
+mod media;
+pub(in crate::engine::durable) use media::{DurableIo, File, StoreLock};
 
 #[derive(Debug)]
 struct Persisted {
